+++ web/src/main/kotlin/com/gmail/pzalejko/invoice/core/model/subject/SellerRepository.kt
@@ -2,7 +2,7 @@ package com.gmail.pzalejko.invoice.core.model.subject
 
 interface SellerRepository {
 
-    fun findByAccountId(accountId: Int): InvoiceSeller?
+    fun findByAccountId(accountId: Long): InvoiceSeller?
 
     fun save(seller: InvoiceSeller)
 }
